+++ test/functional/cluster.js
@@ -4,7 +4,10 @@
 /* eslint sort-keys : 0 */
 /* eslint sort-vars : 0 */
 
-const createClient = require('../../dist');
+import chai from 'chai';
+import createClient from '../../src';
+
+const assert = chai.assert;
 
 describe('Functional: cluster', function () {
   // upping default timeout for Travis-CI builds
@@ -20,9 +23,13 @@ describe('Functional: cluster', function () {
     clientOptions['_index'] = index;
     client = createClient(clientOptions);
     client.indices.createIndex(function (err) {
-      assert.ifError(err);
+      if (err) {
+        return done(err);
+      }
       client.cluster.health({ 'wait_for_status': 'yellow' }, function (err) {
-        assert.ifError(err);
+        if (err) {
+          return done(err);
+        }
         done();
       });
     });
@@ -40,7 +47,9 @@ describe('Functional: cluster', function () {
   describe('#fieldStats', function () {
     it('should be able to get field stats', function (done) {
       client.cluster.fieldStats({ field: '*' }, function (err, result) {
-        assert.ifError(err);
+        if (err) {
+          return done(err);
+        }
         assert(result.indices);
         done();
       });
@@ -50,7 +59,9 @@ describe('Functional: cluster', function () {
   describe('#health', function () {
     it('should be able to get the cluster health', function (done) {
       client.cluster.health({ 'wait_for_status': 'yellow' }, function (err, result) {
-        assert.ifError(err);
+        if (err) {
+          return done(err);
+        }
         assert.equal(result.status, 'yellow');
         done();
       });
@@ -59,8 +70,11 @@ describe('Functional: cluster', function () {
 
   describe('#hotThreads', function () {
     it('should be able to get the hot threads', function (done) {
-      client.cluster.hotThreads(function (result) {
-        assert(result.message.indexOf('cpu usage') > 0);
+      client.cluster.hotThreads(function (err, result) {
+        if (err) {
+          return done(err);
+        }
+        assert(result);
         done();
       });
     });
@@ -69,7 +83,9 @@ describe('Functional: cluster', function () {
   describe('#nodesInfo', function () {
     it('should be able to get nodes info', function (done) {
       client.cluster.nodesInfo(function (err, result) {
-        assert.ifError(err);
+        if (err) {
+          return done(err);
+        }
         assert(result.nodes);
         done();
       });
@@ -79,7 +95,9 @@ describe('Functional: cluster', function () {
   describe('#nodesStats', function () {
     it('should be able to get node stats', function (done) {
       client.cluster.nodesStats(function (err, result) {
-        assert.ifError(err);
+        if (err) {
+          return done(err);
+        }
         assert(result.nodes);
         done();
       });
@@ -109,7 +127,9 @@ describe('Functional: cluster', function () {
   describe('#state', function () {
     it('should be able to get the state', function (done) {
       client.cluster.state(function (err, result) {
-        assert.ifError(err);
+        if (err) {
+          return done(err);
+        }
         assert(result.metadata);
         done();
       });